+++ R/oauth-endpoint.r
@@ -1,4 +1,10 @@
+#' @include url.r
+NULL
+
 #' Describe an OAuth endpoint.
+#'
+#' See \code{\link{oauth_endpoints}} for a list of popular OAuth endpoints
+#' baked into httr.
 #'
 #' @param request url used to request initial (unauthenticated) token.
 #'   If using OAuth1.0, leave as \code{NULL}.
@@ -16,6 +22,8 @@
 #' facebook <- oauth_endpoint(
 #'   authorize = "https://www.facebook.com/dialog/oauth",
 #'   access = "https://graph.facebook.com/oauth/access_token")
+#'   
+#' oauth_endpoints
 oauth_endpoint <- function(request = NULL, authorize, access, base_url = NULL) {
   if (is.null(base_url)) {
     return(endpoint(request, authorize, access))
@@ -34,7 +42,7 @@ endpoint <- function(request, authorize, access) {
     class = "oauth_endpoint")
 }
 
-is.oauth_endpoint <- function(x) inherits(x, "endpoint")
+is.oauth_endpoint <- function(x) inherits(x, "oauth_endpoint")
 
 #' @export
 print.oauth_endpoint <- function(x, ...) {
@@ -43,3 +51,24 @@ print.oauth_endpoint <- function(x, ...) {
   cat("  authorize: ", x$authorize, "\n", sep = "")
   cat("  access:    ", x$access, "\n", sep = "")
 }  
+
+#' Popular oauth endpoints.
+#' 
+#' This list provides some common OAuth endpoints.
+#' 
+#' @export
+oauth_endpoints <- list(
+  linkedin = oauth_endpoint(base_url = "https://api.linkedin.com/uas/oauth",
+    "requestToken", "authorize", "accessToken"),
+  twitter = oauth_endpoint(base_url = "http://api.twitter.com/oauth",
+    "request_token", "authenticate", "access_token"),
+  vimeo = oauth_endpoint(base_url = "http://vimeo.com/oauth",
+    "request_token", "authorize", "access_token"),
+  google = oauth_endpoint(base_url = "https://accounts.google.com/o/oauth2",
+    NULL, "auth", "token"),
+  facebook = oauth_endpoint(
+    authorize = "https://www.facebook.com/dialog/oauth",
+    access = "https://graph.facebook.com/oauth/access_token"),
+  github = oauth_endpoint(base_url = "https://github.com/login/oauth",
+    NULL, "authorize", "access_token")
+)