+++ demo/oauth2-facebook.r
@@ -2,9 +2,7 @@ library(httr)
 
 # 1. Find OAuth settings for facebook:
 #    http://developers.facebook.com/docs/authentication/server-side/
-facebook <- oauth_endpoint(
-  authorize = "https://www.facebook.com/dialog/oauth",
-  access = "https://graph.facebook.com/oauth/access_token")
+oauth_endpoints$facebook
 
 # 2. Register an application at https://developers.facebook.com/apps/
 #    Insert your values below - if secret is omitted, it will look it up in
@@ -12,5 +10,5 @@ facebook <- oauth_endpoint(
 myapp <- oauth_app("facebook", "353609681364760")
 
 # 3. Get OAuth credentials
-facebook_token <- oauth2.0_token(facebook, myapp,
+facebook_token <- oauth2.0_token(oauth_endpoints$facebook, myapp,
   type = "application/x-www-form-urlencoded")